backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic-settings==2.6.1
watchdog==6.0.0
pillow==11.0.0
imagehash==4.3.1
httpx==0.28.1
rapidocr-onnxruntime==1.4.4
numpy==2.4.4
python-multipart==0.0.20
